move getGeoJson into FaultModelControls; simplify reducer;
